+++ src/setup_payload/QRCodeSetupPayloadParser.cpp
@@ -46,8 +46,8 @@ static CHIP_ERROR readBits(std::vector<uint8_t> buf, size_t & index, uint64_t & 
     dest = 0;
     if (index + numberOfBitsToRead > buf.size() * 8 || numberOfBitsToRead > sizeof(uint64_t) * 8)
     {
-        ChipLogError(SetupPayload, "Error parsing QR code. startIndex %zu numberOfBitsToLoad %zu buf_len %zu ", index,
-                     numberOfBitsToRead, buf.size());
+        ChipLogError(SetupPayload, "Error parsing QR code. startIndex %u numberOfBitsToLoad %u buf_len %u ", (unsigned int) index,
+                     (unsigned int) numberOfBitsToRead, (unsigned int) (buf.size()));
         return CHIP_ERROR_INVALID_ARGUMENT;
     }
